Moved vosk model getting code to vosk_local
