+++ src/cli.rs
@@ -1,9 +1,5 @@
 use clap::Parser;
-use reqwest::blocking::Client;
-use std::fs::File;
-use std::io::{self, Cursor};
 use std::path::Path;
-use zip::read::ZipArchive;
 use dirs::home_dir;
 
 #[derive(Parser)]
@@ -84,97 +80,3 @@ fn thread_number_in_range(t: &str) -> Result<usize, String> {
     }
 }
 
-pub fn get_model(model: &str, model_path: String) {
-    let url = match model {
-        "small" => "https://alphacephei.com/vosk/models/vosk-model-small-en-us-0.15.zip",
-        "medium" => "https://alphacephei.com/vosk/models/vosk-model-en-us-0.22-lgraph.zip",
-        "large" => "https://alphacephei.com/vosk/models/vosk-model-en-us-0.22.zip",
-        _ => panic!("It should be impossible to get here"),
-    };
-
-    let output_dir = Path::new(&model_path);
-
-    if output_dir.exists() {
-        std::fs::remove_dir_all(output_dir).expect("Error removing current model");
-    }
-
-    println!("Getting model '{}' at {}", model, url);
-
-    // Download the ZIP file
-    let zip_data = download_file(url).expect(&format!("Error downloading file: {}", url));
-
-    // Unzip the downloaded file
-    unzip_file(&zip_data, output_dir).expect("Error unzipping file");
-
-    let mut entry =
-        std::fs::read_dir(output_dir).expect("error getting downloaded model directory");
-
-    let model_dir = entry.next().unwrap().unwrap();
-
-    unwrap_model(&model_dir.path()).expect("Error unwrapping model directories");
-
-}
-
-fn download_file(url: &str) -> Result<Vec<u8>, reqwest::Error> {
-    let client = Client::builder()
-        .timeout(std::time::Duration::from_secs(600)) // Set a timeout of 10 minutes
-        .build()?;
-
-    let mut response = client.get(url).send()?;
-    let mut content = Vec::new();
-    response.copy_to(&mut content)?;
-
-    println!("Done downloading model");
-
-    Ok(content)
-}
-
-fn unzip_file(data: &[u8], output_dir: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let reader = Cursor::new(data);
-    let mut zip = ZipArchive::new(reader)?;
-
-    for i in 0..zip.len() {
-        let mut file = zip.by_index(i)?;
-        let outpath = match file.enclosed_name() {
-            Some(path) => output_dir.join(path),
-            None => continue,
-        };
-
-        if file.name().ends_with('/') {
-            std::fs::create_dir_all(&outpath)?;
-        } else {
-            if let Some(p) = outpath.parent() {
-                if !p.exists() {
-                    std::fs::create_dir_all(p)?;
-                }
-            }
-            let mut outfile = File::create(&outpath)?;
-            io::copy(&mut file, &mut outfile)?;
-        }
-    }
-    println!("Done unzipping model");
-    Ok(())
-}
-
-fn unwrap_model(model_dir: &Path) -> Result<(), Box<dyn std::error::Error>> {
-    let parent_folder = model_dir.parent().expect("Folder has no parent");
-
-    // Read the contents of the folder
-    let entries = std::fs::read_dir(model_dir)?;
-
-    for entry in entries {
-        let entry = entry?;
-        let entry_path = entry.path();
-        let file_name = entry_path.file_name().expect("Entry has no file name");
-        let new_path = parent_folder.join(file_name);
-
-        // Move the entry to the parent folder
-        std::fs::rename(entry_path, new_path)?;
-    }
-
-    std::fs::remove_dir(model_dir)?;
-
-    println!("Done unwrapping model");
-
-    Ok(())
-}
